refactor(cli): migrate ecmwf tooling onto discovery — close #863
Move ECMWF's refresher/writer/coverage/prober/deep_prober/emitter/live_validator
plus the categoriser, hydrator and seeder to earthlens.ecmwf.cli, relocating the
_categories / _hydrate / _seed helper modules into the provider distribution.
Generalise core's datasets --all seeder and --fill-empty hydrator, and stanza's
write_stanza categoriser, to pure role dispatch (core names no backend). Empty both
guard allow-lists — core's CLI is now fully backend-agnostic, every provider's
catalog tooling discovered via the earthlens.cli entry-point group.
