+++ libs/providers/atmosphere/src/earthlens/_atmosphere_cli/__init__.py
@@ -24,6 +24,18 @@ CLI_TOOLING: dict[str, dict[str, str]] = {
         "bundled_ids": "earthlens.chc.cli:bundled_ids",
     },
     "drought": {"validator": "earthlens.drought.cli:validator"},
+    "ecmwf": {
+        "refresher": "earthlens.ecmwf.cli:refresher",
+        "writer": "earthlens.ecmwf.cli:writer",
+        "coverage": "earthlens.ecmwf.cli:coverage",
+        "prober": "earthlens.ecmwf.cli:prober",
+        "deep_prober": "earthlens.ecmwf.cli:deep_prober",
+        "emitter": "earthlens.ecmwf.cli:emitter",
+        "live_validator": "earthlens.ecmwf.cli:live_validator",
+        "categoriser": "earthlens.ecmwf.cli:categoriser",
+        "hydrator": "earthlens.ecmwf.cli:hydrator",
+        "seeder": "earthlens.ecmwf.cli:seeder",
+    },
     "goes": {"validator": "earthlens.goes.cli:validator"},
     "mswep": {"validator": "earthlens.mswep.cli:validator"},
     "nrel": {"validator": "earthlens.nrel.cli:validator"},
