+++ libs/providers/atmosphere/src/earthlens/ecmwf/_categories.py
